refactor(staking): remove common types file
